Moves Jira authentication to the superclass GLVSC-554
