+++ src/plus/integrations/authentication/gitlab.ts
@@ -1,16 +1,10 @@
 import type { AuthenticationSession, Disposable, QuickInputButton } from 'vscode';
 import { env, ThemeIcon, Uri, window } from 'vscode';
-import type {
-	IntegrationAuthenticationProvider,
-	IntegrationAuthenticationSessionDescriptor,
-} from './integrationAuthentication';
+import type { IntegrationAuthenticationSessionDescriptor } from './integrationAuthentication';
+import { IntegrationAuthenticationProvider } from './integrationAuthentication';
 
-export class GitLabAuthenticationProvider implements IntegrationAuthenticationProvider {
-	getSessionId(descriptor?: IntegrationAuthenticationSessionDescriptor): string {
-		return descriptor?.domain ?? '';
-	}
-
-	async createSession(
+export class GitLabAuthenticationProvider extends IntegrationAuthenticationProvider {
+	override async createSession(
 		descriptor?: IntegrationAuthenticationSessionDescriptor,
 	): Promise<AuthenticationSession | undefined> {
 		const input = window.createInputBox();
@@ -75,5 +69,9 @@ export class GitLabAuthenticationProvider implements IntegrationAuthenticationPr
 				label: '',
 			},
 		};
+	}
+
+	protected override getCompletionInputTitle(): string {
+		throw new Error('Method not implemented');
 	}
 }
